# openapi/api/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "HealthService"
    },
    {
      "name": "AdminService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/admin/definitions": {
      "get": {
        "operationId": "AdminService_ListDefinitions",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1Definition"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1Definition"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "eager",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "SORT_DIRECTION_UNSPECIFIED",
              "SORT_DIRECTION_ASCENDING",
              "SORT_DIRECTION_DESCENDING"
            ],
            "default": "SORT_DIRECTION_UNSPECIFIED"
          },
          {
            "name": "afterKey",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "maxResults",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "AdminService"
        ]
      },
      "post": {
        "summary": "CreateDefinition creates a new definition object.",
        "operationId": "AdminService_CreateDefinition",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateDefinitionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateDefinitionRequest"
            }
          }
        ],
        "tags": [
          "AdminService"
        ]
      }
    },
    "/v1/admin/definitions/{id}": {
      "get": {
        "operationId": "AdminService_GetDefinition",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetDefinitionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "AdminService"
        ]
      }
    },
    "/v1/admin/fields": {
      "post": {
        "operationId": "AdminService_CreateField",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateFieldResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateFieldRequest"
            }
          }
        ],
        "tags": [
          "AdminService"
        ]
      }
    },
    "/v1/health/ping": {
      "get": {
        "summary": "Ping is used to test that the service is up and responding to requests.",
        "operationId": "HealthService_Ping",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1PingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "HealthService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateDefinitionRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Field"
          }
        }
      }
    },
    "v1CreateDefinitionResponse": {
      "type": "object",
      "properties": {
        "definition": {
          "$ref": "#/definitions/v1Definition"
        }
      }
    },
    "v1CreateFieldRequest": {
      "type": "object",
      "properties": {
        "definitionId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "fieldType": {
          "$ref": "#/definitions/v1FieldType"
        },
        "constraints": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1CreateFieldResponse": {
      "type": "object",
      "properties": {
        "field": {
          "$ref": "#/definitions/v1Field"
        }
      }
    },
    "v1Definition": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Field"
          }
        }
      }
    },
    "v1Field": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "fieldType": {
          "$ref": "#/definitions/v1FieldType"
        },
        "constraints": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "v1FieldType": {
      "type": "string",
      "enum": [
        "FIELD_TYPE_UNSPECIFIED",
        "FIELD_TYPE_FLOAT",
        "FIELD_TYPE_INT",
        "FIELD_TYPE_STRING",
        "FIELD_TYPE_TEXT"
      ],
      "default": "FIELD_TYPE_UNSPECIFIED"
    },
    "v1GetDefinitionResponse": {
      "type": "object",
      "properties": {
        "definition": {
          "$ref": "#/definitions/v1Definition"
        }
      }
    },
    "v1PingResponse": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        }
      }
    },
    "v1SortDirection": {
      "type": "string",
      "enum": [
        "SORT_DIRECTION_UNSPECIFIED",
        "SORT_DIRECTION_ASCENDING",
        "SORT_DIRECTION_DESCENDING"
      ],
      "default": "SORT_DIRECTION_UNSPECIFIED"
    }
  }
}